requirements.txt
Flask==2.2.5
Flask-CORS==3.0.10
python-dotenv==0.21.1
scikit-learn==1.0.2
numpy==1.23.5
pandas==1.5.3
